+++ src/components/Header/Header.jsx
@@ -1,16 +1,14 @@
 import { useState } from 'react';
-import { FaUser } from 'react-icons/fa';
 import { Link } from 'react-router-dom';
 
 import MenuIcon from '@mui/icons-material/Menu';
 import IconButton from '@mui/material/IconButton';
-import Menu from '@mui/material/Menu';
-import MenuItem from '@mui/material/MenuItem';
 import SwipeableDrawer from '@mui/material/SwipeableDrawer';
 import Toolbar from '@mui/material/Toolbar';
 
 import { route } from '@/constants/route.js';
 import { NavigateMenu } from '../NavigateMenu/NavigateMenu.jsx';
+import { NotAuthUserMenu } from '../NotAuthUserMenu/NotAuthUserMenu.jsx';
 import { UserMenu } from '../UserMenu/UserMenu.jsx';
 import { AppBarStyled, AppContainerStyled } from './Header.styled.js';
 
@@ -21,13 +19,9 @@ import Logo from '../../assets/images/logo.svg';
 const user = { isAuth: true, roles: ['user', 'chef', 'admin', ''] };
 
 const Header = () => {
-  const [anchorEl, setAnchorEl] = useState(null);
   const [state, setState] = useState({
     left: false,
   });
-  const handleMenuOpen = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
 
   const toggleDrawer = (open) => (event) => {
     if (
@@ -41,10 +35,6 @@ const Header = () => {
     setState({ left: open });
   };
 
-  const handleMenuClose = () => {
-    setAnchorEl(null);
-  };
-
   return (
     <AppBarStyled position="static">
       <AppContainerStyled>
@@ -53,29 +43,7 @@ const Header = () => {
         </Link>
 
         <Toolbar>
-          {!user.isAuth && (
-            <>
-              <IconButton onClick={handleMenuOpen}>
-                <FaUser />
-              </IconButton>
-              <Menu
-                anchorEl={anchorEl}
-                open={Boolean(anchorEl)}
-                onClose={handleMenuClose}
-              >
-                {!user.isAuth && (
-                  <MenuItem onClick={handleMenuClose}>
-                    <Link to={route.SIGN_UP}>Sign up</Link>
-                  </MenuItem>
-                )}
-                {!user.isAuth && (
-                  <MenuItem onClick={handleMenuClose}>
-                    <Link to={route.SIGN_IN}>Sign in</Link>
-                  </MenuItem>
-                )}
-              </Menu>
-            </>
-          )}
+          {!user.isAuth && <NotAuthUserMenu />}
           {user.isAuth && <UserMenu />}
 
           <IconButton aria-label="menu" onClick={toggleDrawer(true)}>
